# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API сервис для автоматизации рекрутинга.",
        "title": "AI Recruiting API",
        "contact": {},
        "version": "1.0"
    },
    "host": "ai-recruiting.onrender.com",
    "basePath": "/",
    "paths": {
...
        "/templates": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Список всех шаблонов рекрутера",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID рекрутера",
                        "name": "recruiter_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.MessageTemplate"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Создать шаблон сообщения",
                "parameters": [
                    {
                        "description": "Данные шаблона",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "body_text": {
                                    "type": "string"
                                },
                                "recruiter_id": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/templates/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "templates"
                ],
                "summary": "Обновить шаблон",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "body_text": {
                                    "type": "string"
                                },
                                "title": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "templates"
                ],
                "summary": "Удалить шаблон",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID шаблона",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/vacancies": {
            "get": {
                "description": "Получает все активные вакансии",
                "tags": [
                    "vacancies"
                ],
                "summary": "Список активных вакансий",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/models.Vacancy"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Создает вакансию и автоматически генерирует short_link",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vacancies"
                ],
                "summary": "Создание вакансии",
                "parameters": [
                    {
                        "description": "Объект вакансии",
                        "name": "vacancy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VacancyCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Вакансия успешно создана"
                    },
                    "400": {
                        "description": "Ошибка парсинга JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/vacancies/{id}/applications": {
            "get": {
                "description": "Получает все заявки кандидатов для конкретной вакансии",
                "tags": [
                    "vacancies"
                ],
                "summary": "Список откликов на вакансию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID вакансии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Application"
                            }
                        }
                    }
                }
            }
        },
        "/vacancies/{id}/archive": {
            "patch": {
                "description": "Переводит вакансию в статус архивной по ID",
                "tags": [
                    "vacancies"
                ],
                "summary": "Архивация вакансии",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID вакансии",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacancy archived",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Вакансия не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Application": {
            "type": "object",
            "properties": {
                "ai_score": {
                    "type": "integer"
                },
                "applied_at": {
                    "type": "string"
                },
                "candidate_id": {
                    "type": "string"
                },
                "candidate_name": {
                    "description": "Для отображения в списке рекрутера",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "vacancy_id": {
                    "type": "string"
                }
            }
        },
        "models.MessageTemplate": {
            "type": "object",
            "properties": {
                "body_text": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "recruiter_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Vacancy": {
            "type": "object",
            "properties": {
                "ai_filters": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_archived": {
                    "type": "boolean"
                },
                "recruiter_id": {
                    "type": "string"
                },
                "short_link": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.VacancyCreateRequest": {
            "type": "object",
            "properties": {
                "ai_filters": {
                    "type": "string"
                },
                "is_archived": {
                    "type": "boolean"
                },
                "recruiter_id": {
                    "type": "string"
                },
                "short_link": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}